aoc21: Fix path and cython compilation
Path needs to be from the parent folder of the day/ themselves. This is
to make it easier to execute the files using '2021/run.py' and it allows
autocompletion of the parameter to the script like the input file to
use.

Cython does not support __future__ and declaration of the type of object
that is passed to a method within itself (recursiveness)